No longer using vectorized singleCutHist in nls
This was the biggest (only) bottleneck in the analysis. By capturing the
raw single cut values as an element in the flowHist object, we can then
multiply this vector by SCa in the nls model, instead of recalculating
the vectorized values. This is much, much faster. We do, however, lose
the ability to use `integrate` to calculate the total number of modelled
events. We still have the number of events modelled in each peak, which
are the values we actually need.