Added linearity option
Controlled with a new slot value in the FlowHist object. Updated the
ModelComponent object to allow for dynamically creating specialParams
lists, providing a mechanism for each model component to select which
params are special based on properties in the parent FlowHist object.

Also updated resetFlowHist option to allow it to keep the peaks when
reseting other slots.