Added third gaussian comp, S-phase comps now include linearity
Added code to model a third sample, to account for cases where two separate
standards are include in the same sample. Also reorganized the file so that
all components for each sample are together (G1, G2 and S-phase), making it
easier to add a fourth, fifth etc component if necessary. This now involves
copying and pasting large sections of code, which with some thought could
be generalized.

Also fixed a bug in the way the broadened rectangles were bounded.
Originally, they were always bounded by the mean of the G1 peak and the
2x the mean of the G1 peak. This assumes the G2 peak is 2x the G1 peak,
which is not true if linearity = "variable". To correct this, added d as
a (special) parameter of the broadened rectangle components.