Integrated gating UI into browseFlowHist
Changed from sidePanelLayout to regular fluidRow, and added a second
fluidRow beneath to hold the gating UI. Basic workflow works, but need a
way to visualize existing gates. As is, existing gates are incorporated in
analysis, but cannot be examined - you can either set a new gate, or accept
the unseen old gate.