Added an action button for setting/unsetting the gate
Shiny doesn't provide a way to control the brushPoints in a fine-grained
way. So now we use a two-stage process - select the gate, and see a preview
in the small plots. When you like what you see, click 'set gate' to apply
the gate to the real data.

To do: sort out updating the gate plot after, not before, the new gate is set.