Completed browseFlowHist reorganization
fhPlot removed, replaced by a suite of reactives that each handle a single
operation. This allows for direct, fine-grained control of the reaction
chain, and everything seems to be working as intended. I think this
reorganization even reduces (some of) the redundant re-analysis of FlowHist
objects that I hadn't been able to eliminate before.